[megaco|test] More test case tweaking
A test case failed (on a slooow linux box) because things
seem to just stop (no action for approx 60 sec on one node
and 10 sec on another) for no apparent reason. And this time
there where no system events to explain it.
It "seemed" to happen in a place in the code where a
   "receive ... after 0 ... end"
was executed.