beam_kernel_to_ssa: Try to harder to avoid try/catch
When evaluating a guard expression in a try/catch, try harder
to eliminate the try/catch.

This is important for expression in binaries and maps in a receive.
In a receive, it is not allowed to use a try/catch, because it will
force a stack frame and saving of registers to the stack frame.