ssl: Fix truncate_client_hello/1
truncate_client_hello/1 fails to remove the binders if a binder with a
stronger hash algorithm (e.g. sha384) ends with a binary that matches
a shorter binder with a weaker hash (e.g. sha256):

opaque PskBinderEntry<32..255>;

 48
126  88  24 102 125 162 198  13
 64  74  59   9 130 160   1  32*
 77  25 108 227 168  11 145  83
110 132 159 135  67 195 186  97
 54  25 248 131 234 187  58 211
159 120  48 198  17  76 242 138

This commit fixes the truncate_client_hello/1 function by parsing
the ClientHello message and setting binders to an empty list.