cover: Suppress compilation warnings when instrumenting andalso/orelse
The cover tool translates andalso/orelse to a case in the instrumented
code. For example:

    both_ok(A, B) ->
	A =:= ok andalso B =:= ok.

Is translated to:

    both_ok(A, B) ->
	case A =:= ok of
	    true  -> B =:= ok;
	    false -> false;
	    _1    -> error({badarg, _1})
	end.

Since the expression `A =:= ok` clearly is a boolean expression, the
final clause can't possibly match, so the compiler issues a
warning. If the `warnings_as_errors` option is set in the environment
or in the module itself, compilation will fail.

Eliminate the warning by letting cover mark the final clause as compiler
generated to suppress the warning.

https://bugs.erlang.org/browse/ERL-1147