Add filelib:safe_relative_path/2
Deprecate filename:safe_relative_path/1 in favor of
filelib:safe_relative_path/2.

The filename function does not consider symlinks so it is unsafe to use
on any filesystem that supports symlinks. The new filelib function
follows symlinks and ensures they do not escape the root of the relative
path. It also detects loops in the symlinks to prevent DOS attacks.