[snmp|test] End per testcase tweaking and crypto start
In the end_per_testcase function the logs generated during the
test case is displayed/printed. But this can on some platforms
take a long time, which can result in an otherwise successful
test case to be reported as failed (timeout).
So, set a successful test case).

Also, try to be a bit less verbose when skipping if crypto
is not available.