[snmp|agent|test] Tweaked the snmp-framework-mib test case(s)
The snmp-framework-mib test case(s) tested, among other things,
the engine-time. The test case paused (sleep) for 5 seconds,
and assumed that that was exactly the time it slept. But
that was not always the case. It could be a lot longer.
So, instead we need to use the time we actually slept in the
test.