beam_ssa_type: Very slightly optimize some switches
The loader will choose between different instructions for a
select_val instruction. If the values are integers in a compact
range, a jump table will be used. There are different instructions
for jump tables starting with the value 0 and starting with other
values.

To ensure that the shorter jump table instruction can be used, when
the failure label for the switch can't be reached, be sure to retain
0 as the first choice in the switch.