[snmp|manager|test] Moved manager start and stop into pre and post
In the usm_priv_aes test case, moved the snmp manager start and stop
into the pre and post fun:s, respectively.
The reason for this is that if the test case fails badly, it will
crash out before the manager has been stop'ed, leaving a running
manager. Which will cause problems for later test cases.
This problem is not unique to this test case! Many test cases to
this! We should really fix this *everywhere*...

OTP-16541