Tweak traffic counters for remote request handler processes
Redo the solution in commit d1838e6b to spawn a process on the transport
node instead of banging the diameter_peer_fsm process. Maybe a little
more costly, but banging introduces an upgrade problem: nodes on which
transport is terminated have to be upgraded before nodes handling
request or else the receiving diameter_peer_fsm process can crash on an
unexpected message, taking a peer connection with it.

Not an odious requirement as long as transport/handler nodes are
distinct, but it still has to be documented wheres spawning into an
existing function just works. In the case that there isn't a clean
transport/handler division, the solution would be to stop and then start
all nodes in the cluster, or just accept that peer connections can be
lost, but not having a clear division is probably questionable use of
remote handlers: the point is to be able to move request handling from a
node that terminates transport to another node with less to do, not to
another node that terminates transport.

This also has the advantage of counters not being dropped if the fsm
process goes down. The handler process increments them independently of
the fsm process just as if it had been on the transport node.

Configure {traffic_counters, false} to skip counters altogether, which
is the cheapest solution.