v3_core: Eliminate crash when clauses can't match
The compiler could crash if clauses that couldn't possibly match were
removed because of variables no longer being bound. Here is an
example:

    foobar() ->
	[receive
	    a = <<V0>> ->
		 any
	 end],
	V0.

An attempt to fix this problem for receive was made in 051ecbc1b3b7,
but the fix was not general enough.

This commit avoids the problem by never removing clauses that can't
match. Instead, the pattern is rewritten so that it binds the same
variables and the guard is changed to `false`.