sys_core_fold: Optimize boolean cases more conservatively
When optimizing a case construct, if the case expression is a boolean
expression that is not guaranteed to return (it calls erlang:error/1),
don't do any boolean optimizations. The boolean optimizations will be
done anyway by beam_ssa_type. If we would do the boolean optimizations
anyway, the result would be a case where the last clause is not
guaranteed to match, and that can potentially confuse other passes.