v3_core: Don't make the compiler crash on invalid map updates
Consider this code that does a map update on the integer 42:

    42 #{part => V = false},
    V

Since 42 is not a map, v3_core would rewrite the code like this:

    erlang:error(badmap),
    V

In this particular example, since `V` is clearly unreachable,
sys_core_fold will remove the use of `V` and the compiler will not
crash. However, the Core Linter will complain if it is enabled.

The compiler will crash when given some more complicated examples:

    foobar() ->
	case 42 of
	    V ->
		[];
	    power ->
		[]#{key =>
			receive
			    V -> false
			end}
	end,
	V.

To avoid this problem, stop rewriting failing map updates to
a call to `error/1`.