Normalize terminators to eliminate compiler crash
Code such as the following could crash the compiler if optimizing
passes were disabled:

    if
        0.1 orelse "VZ", 42 -> ok;
        true -> error
    end

The crash was caused by denormalized terminators, such as:

    bool = bif:'=:=' ...
    br bool, ^2, ^2

Here the terminator references a boolean variable, but the success
and failure branches are the same. That will cause the `bool` variable
to be allocated to z register, and the z register will not be eliminated
by beam_ssa_codegen. Eliminate the issue by normalizing terminators
in all compiler passes that updates a a terminator. For this example,
the normalized terminators looks like this:

    bool = bif:'=:=' ...
    br ^2

A similar problem for code generation is a succeeded:guard instruction
not followed by a two-way branch:

    result = <some instruction>
    did_succeed = succeeded:guard result
    br ^99

Eliminate this problem by making sure that any pass that changes a
terminator into a one-way branch also removes any preceeding
succeeded:guard and the instruction before it.