[erts] Tweaked the config and usage of SCTP for NetBSD
On NetBSD the type sctp_assoc_value does not exist.
Its used normally when setting, for instance, the
option SCTP_DELAYED_ACK_TIME. This option *does*
exist on NetBSD, but it can only be used to set the
"default value" (which on Linux is done by setting
assoc_id (in assoc_value) to 0).
The way to "properly" handle this on NetBSD would be
to accept this option (for inet:[setopts/getopts]),
but verify that the assoc id is zero (0).
For now we simply test for the type, and if it does
not exist, we  do not allow set/get of this option.