[snmp|agent|test] Tweaked test suite to work on NetBSD
We skip all IPv6 test case(s) oín the agent suite on
NetBSD. UDP on IPv6 is glitchy on the only NetBSD machine.
It could be because of configuration, but since the socket
suite works...anyway, its not a snmp problem, so skip
until the root cause is solved.

Improved handlng of udp error(s) in the test manager (used in
the agent suite). If sunch an error occurs, we exit. But since
its not an snmp error we don not fail the test case. Instead
we skip.