beam_ssa_type: Discard unshared variables in join_types/2
This greatly improves performance when a lot of different blocks
with lots of variables individually jump to a single successor,
such as in ERL-1185. It also improves type inference for phi nodes
since we can refer to the types leaving the predecessor blocks
rather than the joined types in the current block.