beam_ssa_lint: Harden invariants
Disallow problematic constructs that can confuse beam_ssa_pre_codegen
and/or beam_ssa_codegen and cause the compiler to crash.

The first construct to forbid is denormalized terminators such as:

    br bool, ^99, ^99

will not be allowed. This terminator must be normalized to:

    br ^99

The second construct is succeeded:guard not followed by two-way
branch:

   previous_result = ...
   succ_bool = succeeded:guard previous_result
   br ^99

If an optimization pass rewrites a two-way to a one-way branch, it
must also remove any preceding succeeded:guard instruction and
the instruction before succeeded:guard.

It is important that every pass respects the invariants, because
it should be possible to skip any optimization pass and the compiler
should still generate code.