[megaco] Updated with regard to deprecated encoding config
The codec(s) prev3a, prev3b and prev3c has been deprecated.
Since they are never called directly by a user, there use
is dependent on encodeding config, which the user provides,
the reprecated attribute is not useful. Instead we rely
on documentation (basically) and comments in the code for
the eventual removal (in OTP 24).
Also, a warning message is issued whenever an encoding config
using the deprecated config is set:

      {version3, prev3a | prev3b | prev3c}

OTP-16531