scripts/diffable: Add option for compiling with +deterministic
This patch adds a new `--deterministic` command line flag to the
diffable script. When enabled, diffable compiles files with the
`deterministic` compiler option enabled. As deterministic compilation
results in line directives only containing the basename of module
files, this greatly reduces the size of diffs when comparing the
results from different build and source trees.