beam_ssa_codegen: Simplify the prefer_xregs/2 optimization
18dc9a8099d1 introduced the `swap` instruction. Before we had
the `swap` instruction it was not always benefical to use
an X register instead of a Y register when a value was available
in both, because we might have to swap two X registers using a
sequence of three `move` instructions.

Simplify the prefer_xregs/2 optimization to unconditionally
prefer X registers over Y registers.

While at it, also improve the code generation for fun calls
and applies.