beam_validator: Split validation into a stronger and a weaker pass
It's not always easy to accurately reconstruct type information
after all assembly-level optimizations have been applied, and with
the OTP 23 release around the corner it's a bit late to perform
major surgery on the validator or type representation.

Therefore, we've decided to split validation into two passses; the
first runs directly after code generation and is as strong as
before, and a second weaker pass runs after all assembly-level
optimizations have run. We'll have to do something about this in
OTP 24, but it will have to do for now.