Properly represent wait_timeout as a three-way branch in SSA
The `wait_timeout` instruction is a three-way branch because it will
raise an exception if the timeout value is invalid. However, in SSA
code `wait_timeout` is represented as a two-way branch (ignoring the
possibility of an exception). Because of the missing edge to the block
with a `landingpad` instruction, unsafe code could be generated for a
`receive after` within a try/catch.

Change the representation of `wait_timeout` in the SSA code to
use a `succeeded:body` instruction to explicitly represent the
exception edge.