feat(socket): pass map as credentials
This is needed due to fact that Erlang do not provide any way to check
size of `pid_t`, `uid_t`, and `gid_t` which is needed for proper
handling of the data passed to the socket. This allows user to pass map
containing 3 fields - `pid`, `uid`, and `gid` that will be changed to
the `struct ucred` before passing via socket. It also decodes data that
come out of the socket when possible.