beam_ssa_codegen: Delay fix ups of wait_timeout
1efd0838fc9b changed the representation of wait_timeout to be
a three-way branch. In that commit, the three-way branch is translated
back to a two-branch early in the beam_ssa_codegen. However, that
can lead to Y registers being killed prematurely. Translating to the
two-way branching instruction must be done after all anlyses are done
and all annotations are in place.