erts: Fix external fun application of erlang:apply/3
    Fun = fun erlang:apply/3,
    ... snip ...
    Fun(erlang, self, []).

Code like the above would always crash because the CIF was just a
stub that threw `badarg`, and the actual implementation in the
`erlang` module was overwritten after an apply/3-specific loader
hack was removed when refactoring CIF tracing.

Since the CIF was only used for its MFA in an error handling macro,
it's better to declare the MFA in said macro and skip the CIF
altogether.