[esock] Comment out the use of PACKET_FASTROUTE
On (at least some versions of) Android the use of this
constant results in compilation errors as it has the
same value as PACKET_USER.
On other linux variants (Linux Mint 19.3) there is a
comment in the include file that seems to indicate
that its un used and also non indented for user land.
But it still compils there, so there must be some undef
somehere (which is missing on Android).

OTP-16576
ERRL-1208