hipe: Use portable comment syntax in assembler files
Ancient versions of GAS (such as the one included in `gcc 3.4.3`)
sometimes confused `#`-comments for pre-processor macros, breaking
the build. Using `//` and `/**/` fixes this.

This may seem a bit silly but modern big-endian machines are very
expensive, so most of our big-endian rigs are ancient, and we have
to adapt if want to test HiPE on them.