beam_trim: Trim the stack more aggressively
The stack trimming used to be very conservative, avoiding stack
trimming if the trimming instruction sequence was estimated to be
slower than the original sequence. That could make recursive functions
using a huge amount of stack slower if unused stack slots were kept.

To avoid the cost of not trimming in recursive functions, adjust the
cost calculation formula to trim more often.

This commit is a partial solution to ERL-1216.