crypto: Curves cache
Currently un-supported curves must be removed from the crypto:supports
result.

It is very slow on some machines (no hw-accel?) so we keep a cache once
the list is filtered. If FIPS mode is supported, the list must be changed
each time the mode is changed.

The list is rebuilt:
  - first call to start() or first call of crypto:supports
  - each state change of the FIPS mode