beam_ssa_opt: Make the sinking of get_tuple_element more conservative
Sinking (delaying) the extraction of tuple elements until the elements
are needed is often advantageous. However, in rare circumstances, the
"optimized" code could become much slower. As an example, take the
following function:

    mapfoldl(F, Acc0, [Hd|Tail]) ->
	{R,Acc1} = F(Hd, Acc0),
	{Rs,Acc2} = mapfoldl(F, Acc1, Tail),
	{[R|Rs],Acc2};
    mapfoldl(F, Acc, []) ->
	{[],Acc}.

If the compiler delays the extraction of tuple elements as long as
possible, the resulting code will be similar to the following:

    slow_mapfoldl(F, Acc0, [Hd|Tail]) ->
	Res1 = F(Hd, Acc0),
	Res2 = slow_mapfoldl(F, element(2, Res1), Tail),
	{[element(1, Res1)|element(1, Res2)],element(2, Res2)};
    slow_mapfoldl(F, Acc, []) ->
	{[],Acc}.

Note that the tuple bound to the `Res1` variable will be kept alive
during the recursive call. That means that all intermediate accumulators
will be kept alive until `slow_mapfoldl/3` returns. In this case, it would
clearly be better to extract all tuple elements at once:

    fast_mapfoldl(F, Acc0, [Hd|Tail]) ->
	Res1 = F(Hd, Acc0),
	R = element(1, Res1),
	Res2 = fast_mapfoldl(F, element(2, Res1), Tail),
	{[R|element(1, Res2)],element(2, Res2)};
    fast_mapfoldl(F, Acc, []) ->
	{[],Acc}.

`fast_mapfoldl/3` uses the same amount of stack space as `slow_mapfoldl/3`.
Thus, `slow_mapfoldl/3` has now advantages whatsoever over `fast_mapfoldl/3`.

To ensure that the compiler emits code similar to the `fast_mapfoldl/3`
example, make the sinking of `get_tuple_element` instructions more
conservative. Only sink when there is an advantage in terms of stack space
or if it can be shown that sinking is essentially harmless.

https://bugs.erlang.org/browse/ERL-1216