sys_core_fold: Remove ineffective optimization of try in guards
The opt_guard_try optimization will remove calls to pure BIFs in a `try`
if the outcome of the call does not matter. This optimization was very
rarely successfully applied (only 34 times in the compiler test
suite), and it is likely that the SSA passes will do the optimization
just as well.