[snmp|agent|test] Fixed agent mibs post tc fun
The post fun was supposed to perform a node cleanup,
but assumed the wrong input. So, exec of the fun crashed.
But since the call was catched it was not visible.
The result was the the tmp nodes was left running,
thereby consuming resources of the host. This is not
a problem for new machines, by we have some really old
and slow kit...