[snmp|agent|test] Each test case is run in its own (new) node
There is some previously run test case that can leave
processes running (symbolic store and mib server).
This will then cause problems for most of the test
the testcases in this suite.
So, to avoid this, all test cases (in this suite) are now
run in their own temporary nodes.