erts: Fix int_drv use of type SOCKLEN_T
SOCKLEN_T is used for the last arguments of
    #define sock_getopt   getsockopt
    #define sock_setopt   setsockopt
    #define sock_name     getsockname
    #define sock_peer     getpeername

On unix this is type socklen_t, but on Windows this is type int.

Commit 011954e851d421b882b0b4eaeda04cfc2895d70a
changed SOCKLEN_T from int to size_t which is not correct on Windows.

Kept size_t fallback on non-Windows with a compile #warning,
not sure if this is ever needed/correct.