[megaco] Mini text parser IPv6 address failure
The mini decoder could handle all kinds of IPv6 addresses.
The reason for this was that the som of the chars (a and e)
in the short form was interpreted as 'AddToken' and 'EventsToke'
respectively, which the mini parser could not handle.

ABNF for the mId (only the IPv6 version):

mId             = (( domainAddress / domainName ) [":" portNumber]) /
                  mtpAddress / deviceName
domainAddress   = "[" (IPv4address / IPv6address) "]"
IPv6address     = hexpart [ ":" IPv4address ]
hexpart         = hexseq "::" [ hexseq ] / "::" [ hexseq ] / hexseq
hexseq          = hex4 *( ":" hex4)
hex4            = 1*4HEXDIG
HEXDIG          = ( DIGIT / "A" / "B" / "C" / "D" / "E" / "F" )
DIGIT           = %x30-39         ; 0-9

Example:  [2409:8050:5005:1243:1011::a]

The final 'a' (or 'e') caused it to fail.

OTP-16631