kernel: Fix boot srv test on ubuntu
For some reason ubuntu sets it's own name to be resolved
to 127.0.1.1 and not 127.0.0.1. This messes with this test
where the ip address that the udp messages are sent from
are validated and turn out to be 127.0.0.1. So we set
the ip to be whatever sovrum resolves to and don't
validate the ip of the boot server.