ssl: Check salt size in PSS signatures
OpenSSL crypto lib defines a limit on the size of the random salt
in PSS signatures based on the size of signing RSA key.
If the limit is unchecked, it causes handshake failures when the
configured certificates contain short (e.g. 1024-bit) RSA keys.
For more information see the OpenSSL crypto library
(rsa_pss:c{77,86}).