ssl: Refactor tests
Start refactoring tests with the goal that tests shall use the same
framework code so that we do not have to duplicate handling of for
instance IP-version.

Avoid asserting openssl s_server and openssl s_client as it is
too error prone to parse its output and it is Erlang we are
testing not OpenSSL. Maybe we can find some better assertions later.

Also fix init_per_group in many test suites so that it better uses
the framework.