Allow any element subpatterns in LHS of ++-patterns
There is no actual need to restrict the elements of the LHS list
of a ++-pattern to be characters like $x or integers like 65. The
rewriting by the compiler is the same even if patterns like
`[foo,bar,baz] ++ Rest` or `[{x,42},{y,Y}]++Rest` are allowed.
Dropping this restriction makes the use of ++ in patterns less of
a special case which exists only for strings.