[kernel|tcp|api-test] Provide more info when failing to find unused ip
Use the net:getifaddrs/0 function to provide more info
when we fail to find an unused info (why we did not).
Note that this does not (yet) work on all platforms (Windows),
in which c ase we behave as before.

Also, replace the "fake" success on the connect_timeout test case
with a skip.