[kernel|tcp|misc-test] Adjust iter-max-socks test case
The test case used rpc to run function on slave node.
This has been replaced with a spawn_opt(monitor) call instead.
We then wait for the DOWN message with the result.
Also, adjust the number of tries. The code to figure out how
many times (iter) to do the 'max socks' calculation was/is
to "simple". But, added a case for darwin (same value as for win32).