Clarify how to make contributions
We want it to be as easy as possible to review a pull request without
breaking the possibility to use bisect.

The need for improvement was discovered due to PR-1968 and a first
enhancement proposal made in PR-2652, but we discovered it needed to
be reworded to fulfill our whole requirement.