Implement EEP-54 (extended error information)
This commit is the reference implementation for EEP-54, which proposes a mechanism
for providing more informational messages than simply "bad argument" when a call
to a BIF fails. Here is an example:

    1> element(a, b).
    ** exception error: bad argument
         in function  element/2
            called as element(a,b)
            *** argument 1: not an integer
            *** argument 2: not a tuple

All the details are found in EEP-54, but here are the main points:

* An extension of the format of the call-stack back trace (**stacktrace**) format
  to indicate that there exists extended error information for that call, and
  a convention for how extended error information can be provided.

* A new `erlang:error/3` BIF to allow libraries and applications to
  raise an exception with extended error information.

* New functions `erl_error:format_exception/3` and
  `erl_error:format_exception/4` to allow libraries and application to
  format stacktraces in the same style as the shell.

This implementation of EEP-54 implements extended error information for all
C-implemented BIFs in the `erlang` and `ets` modules.