Change faulty assert checking that port is not dead at reschedule
After scheduled execution of port tasks, a port is rescheduled
if there still remain scheduled port tasks on it. The assert
checked that the port was not in state ERTS_PORT_SFLGS_DEAD (which
is either ERTS_PORT_SFLG_FREE or ERTS_PORT_SFLG_INITIALIZING) when
being rescheduled.

The port can however be set in state ERTS_PORT_SFLG_FREE between the
point where it is determined that it needs to be rescheduled
(finalize_exec()) and the actual rescheduling where the assert is.

This situation is however not an issue. When this happens the port
will be rescheduled and later selected for execution. The first thing
that happens when it begins executing is that it will complete
termination since it is in state ERTS_PORT_SFLG_FREE.

The modified assert now only checks that the port is not in state
ERTS_PORT_SFLG_INITIALIZING.