sasl: Be more resilient of exiting processes
when asking about their status.

If Proc dies during sys:get_status(Proc), then treat that the
same as if it died before the call (noproc).

The case of disconnection is unchanged. Not because I'm convinced
it's the correct thing to do, but rather because I'm not convinced
it's wrong.