Skip diameter_gen_sctp_SUITE
Testcases in this suite have failed for almost 9 years now. Either
there's a reasonable explanation for it or no one is using gen_sctp
enough to notice/care, but the failures are just noise in diameter's
testing.

May only affect the loopback address, which would explain the lack of
interest.