Only compute vars and keys if we have patterns
The goal of the sys_core_alias pass is to optimize
code like this:

    foo({ok, Val}) ->
      {ok, Val}.

into:

    foo({ok, Val} = Var) ->
      Var.

To do so, we need to find all patterns and find all
tuples and lists that may mirror their patterns.

Previously, we would compute the keys for each data
structure even if no pattern was stored. This optimizes
it so we only compute the keys if we have a pattern
so far.