Fix getting data from ssh-agent
Attempting to request a user key from the ssh agent may fail because `unpack/1` assumes
that the received packet contains the full response, but in some cases it doesn't.

This fixes that case by making 2 `gen_tcp:recv/3` calls - First, fetch 4 bytes for
the expected length of the data response. Then request that expected length to get
the full response before decoding.