ssl: Fix handling of options 'verify' and 'verify_fun'
This change fixes a vulnerability of not verifying peer certificates
when option 'verify' is re-configured from verify_none to verify_peer.

The option 'verify' is simulated by the configured 'verify_fun' that
is mostly hidden from the end user. When 'verify' is set to verify_none,
the option 'verify_fun' is also set to a default verify-none-verify_fun
when processing the configuration. If 'verify' is later changed from
verify_none to verify_peer, the 'verify_fun' must also be changed to
undefined.  When 'verify_fun' is set to undefined, public_key's default
verify_fun will be used that performs a full verification.