Fix prio elevation race
A race condition when changing process priority by calling
process_flag(priority, Prio) could cause elevation of priority for a
system task to be ignored. This bug hit if the system task was
scheduled on the process calling process_flag() at the same time as
the priority was changed. The bug is quite harmless and should hit
very seldom if ever.