Inline always required layer in raw_file_io
The raw option is always available to raw_file_io:open/2
but the previous implementation made it look like it was
optional as the other layers. This commit inlines
raw_file_io_raw into raw_file_io so it is clear it always
ends up by calling prim_file:open/2.