Do not consider dotfiles to be an extension
Previously, filename:extension(".gitignore") would
return ".gitignore" as an extension while ".gitignore"
is actually part of the filename. Similarly,
filename:rootname("foo/.gitignore") would return "foo/".

One argument for .gitignore to be treated as filename
is that we should expect this property to hold:

    length(filename:split(Path)) =:=
        length(filename:split(filename:rootname(Path)))

But that was violated for dotfiles prior to this commit.

Another point of discussion is filenames such as "..ext".
This commit takes the approch that the rootname is "."
and the extension is ".ext".