Change application set/unset env timeouts to infinity
There has been a recent report where starting multiple
processes (8) to perform concurrent httpc requests in
a Raspberry Pi Zero would trigger set_env timeouts.

That's because httpc would call crypto which would call
application:set_env with a cache of the existing curvers
and that would make the aplication_controller unable to
reply in time.

One option to address this issue would be to pass the
timeout option when calling set_env from crypto but
reallistically speaking we would most likely just move
the error to next set_env call that may happen.

Given all functions in application already default to
a timeout of infinity, I chose to fix the problem at
the source and set the default timeout to infinity.

This also aligns with more recent trends in Erlang/OTP
that use an infinity timeout as default.