Allow metadata and ignore to be computed in logger funs
This commit allows the logger metadata to be
computed lazily via the logger `msg_fun`.
The `msg_fun` itself can also return `ignore`.

These features are useful when computing the
metadata or if the message should be logged or
not is expensive and you want to do so only
after the log level match.