Robustify an unsafe loader optimization
The loader would rewrite a code sequence such as the following:

    {move,{atom,something},{x,0}}.
    {move,{x,0},{x,0}}.

To:

    {move,{x,0},{x,0}}.

(If `{x,0}` was uninitialized, that could crash the runtime system.)

Change the loader transformation to keep both `move` instructions.

Note that such code sequences are never emitted by the Erlang compiler,
but it could be emitted by an alternative code generator or found in
handwritten or patched BEAM code.

https://bugs.erlang.org/browse/ERL-1344