httpc: fix deadlock during async request when body as a function
For pipeline/keep-alive connections a deadlock can occur when body is
supplied as a function that communicates with the caller of httpc:request/4
during an async request.

The cause is httpc_manager:handle_request/2 should return immediately
but it calls httpc_request:send/4 (via httpc_manager:pipeline_or_keep_alive/3)
which blocks as httpc_request:send/4 is directly called through that codepath.

When the body function blocks waiting for messages from the caller the
outcome is that the call to httpc:request/4 never returns.  This does not
occur for non-pipeline/keep-alive connections (such as HTTP/1.0) as
httpc_manager:handle_request/2 just calls start_handler and does not block.

https://bugs.erlang.org/browse/ERL-1321