Change cmsghdr format for send
Also, rename msghdr to msg and cmsghdr to cmsg where appropriate.

In Erlang, it makes no sense to separate between the content
and the  message header being a struct pointing to the content,
because the content is contained in the Erlang term.

So it is more apropriate to talk about a msg() containing content.